Projeto_Tablet/requirements.txt
Django==2.2
django-cors-headers==2.5.3
djangorestframework==3.9.2
psycopg2==2.8.2
pytz==2019.1
sqlparse==0.3.0
xlwt==1.3.0
